# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/users":{"get":{"operationId":"UsersController_findAll","summary":"유저정보","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/register":{"post":{"operationId":"UsersController_register","summary":"회원가입","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsersDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsersDTO"}}}},"403":{"description":"forbidden"}},"tags":["Users"],"security":[{"bearer":[]}]}}},"info":{"title":"Nestjs Boilerplate App","description":"The N.B.As API Description","version":"1.0","contact":{}},"tags":[{"name":"Users","description":""}],"servers":[{"url":"http://localhost:3001"}],"components":{"schemas":{"User":{"type":"object","properties":{"id":{"type":"number","example":1,"description":"Auto-Created by DB"},"name":{"type":"string","example":"Jeonghun Lee","description":"The name of the user"},"address":{"type":"string","example":"Seoul-si","description":"The address of the user"},"gender":{"type":"string","example":"male","description":"The gender of the user"},"email":{"type":"string","example":"jhlee11@cookapps.com","description":"E-mail"},"password":{"type":"string","example":"passw@rd","description":"password of the user"},"isActive":{"type":"boolean","example":true,"description":""},"roles":{"example":{"standard":"standard","premium":"premium"},"description":"Users Role","type":"array","items":{"type":"string"}}},"required":["id","name","address","gender","email","password","isActive","roles"]},"UsersDTO":{"type":"object","properties":{"email":{"type":"string","example":"jhlee11@cookapps.com","description":"E-mail"},"address":{"type":"string"},"name":{"type":"string","example":"Jeonghun Lee","description":"The name of the user"},"password":{"type":"string","example":"passw@rd","description":"password of the user"}},"required":["email","address","name","password"]}}}}